feat: replace GPL imessage-database with MIT imessage-db + imessage-core
- Create db.rs wrapper around MessageRepository for paginated queries
- Rewrite exporter.rs: remove streaming, use paginated bulk queries
- Replace BubbleComponent/generate_text with typedstream decoder
- Replace Cocoa epoch conversion with pre-converted Unix ms timestamps
- Update all rowids from i32 to i64
- Simplify tapback resolver (reaction types are strings now)
- Update contacts.rs to use rusqlite directly (no imessage-database)
- Update attachment_manager to use path-based methods
- Fix doc-test crate name references
- Zero GPL dependencies remain

🤖 Created together with [Claude Code](https://claude.com/claude-code)

+++ crates/imessage-ndjson-core/src/serialization/chat.rs
@@ -4,7 +4,7 @@ use serde::Serialize;
 #[derive(Debug, Serialize, Clone)]
 pub struct SerializableChatContext {
     /// Chat ID in database
-    pub chat_id: Option<i32>,
+    pub chat_id: Option<i64>,
     /// Chat identifier (phone number, email, or group chat ID)
     pub chat_identifier: String,
     /// Custom display name for the chat
@@ -20,7 +20,7 @@ pub struct SerializableChatContext {
 #[derive(Debug, Serialize, Clone)]
 pub struct SerializableSender {
     /// Handle ID in database
-    pub handle_id: Option<i32>,
+    pub handle_id: Option<i64>,
     /// Contact identifier (phone number or email)
     pub identifier: String,
     /// Resolved contact name
